+++ python/ray/_common/ray_option_utils.py
@@ -318,7 +318,9 @@ def _warn_if_using_deprecated_placement_group(
 
 
 def validate_task_options(
-    options: Dict[str, Any], in_options: bool, function: Optional[Callable] = None
+    options: Dict[str, Any],
+    in_options: bool,
+    is_generator_callable: Optional[bool] = None,
 ):
     """Options check for Ray tasks.
 
@@ -326,9 +328,9 @@ def validate_task_options(
         options: Options for Ray tasks.
         in_options: If True, we are checking the options under the context of
             ".options()".
-        function: Optional function to validate num_returns='streaming' or 'dynamic'
-            against. If provided and num_returns is 'streaming' or 'dynamic',
-            validates that the function is a generator function.
+        is_generator_callable: Optional bool indicating whether the callable is a
+            generator function. If provided and num_returns is 'streaming' or
+            'dynamic', validates that the callable is a generator.
     """
     for k, v in options.items():
         if k not in task_options:
@@ -341,11 +343,10 @@ def validate_task_options(
         raise ValueError("Setting 'max_calls' is not supported in '.options()'.")
     _check_deprecate_placement_group(options)
 
-    # Validate num_returns for generator functions when function is provided
-    if function is not None:
+    if is_generator_callable is not None:
         num_returns = options.get("num_returns")
         if num_returns is not None:
-            validate_num_returns(function, num_returns)
+            validate_num_returns(is_generator_callable, num_returns)
 
 
 def validate_actor_options(options: Dict[str, Any], in_options: bool):
@@ -386,7 +387,7 @@ def validate_actor_options(options: Dict[str, Any], in_options: bool):
     _check_deprecate_placement_group(options)
 
 
-def validate_num_returns(function: Callable, num_returns: Any) -> None:
+def validate_num_returns(is_generator_callable: bool, num_returns: Any) -> None:
     """Validate num_returns for @ray.remote and @ray.method decorators.
 
     This function validates:
@@ -395,12 +396,13 @@ def validate_num_returns(function: Callable, num_returns: Any) -> None:
        function, it should fail fast.
 
     Args:
-        function: The function or method being decorated.
+        is_generator_callable: Whether the callable is a generator function or
+            async generator function.
         num_returns: The num_returns value to validate.
 
     Raises:
         ValueError: If num_returns < 0, or if num_returns is 'streaming' or 'dynamic'
-            but the function is not a generator function or async generator function.
+            but the callable is not a generator function or async generator function.
     """
     if num_returns is None:
         return
@@ -410,16 +412,12 @@ def validate_num_returns(function: Callable, num_returns: Any) -> None:
         raise ValueError(f"num_returns must be >= 0, but got {num_returns}.")
 
     # Validate num_returns='streaming' or 'dynamic' for generator functions
-    if num_returns in ("streaming", "dynamic"):
-        if not (
-            inspect.isgeneratorfunction(function)
-            or inspect.isasyncgenfunction(function)
-        ):
-            raise ValueError(
-                f"num_returns='{num_returns}' can only be used with generator functions "
-                f"(functions that use 'yield'). "
-                f"Function '{function.__name__}' is not a generator function."
-            )
+    if num_returns in ("streaming", "dynamic") and not is_generator_callable:
+        raise ValueError(
+            f"num_returns='{num_returns}' can only be used with generator functions "
+            f"(functions that use 'yield'). "
+            f"The decorated function is not a generator function."
+        )
 
 
 def update_options(
