+++ python/ray/tests/BUILD.bazel
@@ -540,7 +540,6 @@ py_test_module_list(
         "test_local_port_service_discovery.py",
         "test_log_dedup.py",
         "test_memory_scheduling.py",
-        "test_method_num_returns.py",
         "test_metrics_agent_2.py",
         "test_microbenchmarks.py",
         "test_mini.py",
@@ -556,7 +555,6 @@ py_test_module_list(
         "test_reconstruction_stress.py",
         "test_reconstruction_stress_spill.py",
         "test_reference_counting.py",
-        "test_remote_num_returns.py",
         "test_runtime_env_fork_process.py",
         "test_runtime_env_get_wheel_names.py",
         "test_runtime_env_py_executable.py",
