+++ python/ray/tests/unit/test_decorator_validation.py
@@ -1,8 +1,94 @@
-"""Tests for @ray.method decorator num_returns validation."""
+"""Tests for @ray.remote and @ray.method decorator validation."""
 
 import pytest
 
 import ray
+
+
+class TestRemoteNumReturns:
+    """Test num_returns validation for @ray.remote decorator."""
+
+    def test_num_returns_negative_raises_error(self):
+        """Test that num_returns < 0 raises ValueError at decoration time."""
+        # Option validation happens before validate_num_returns, so it raises
+        # a different error message, but still validates that negative values fail fast.
+        with pytest.raises(ValueError, match="non-negative integer"):
+
+            @ray.remote(num_returns=-1)
+            def f():
+                return 1
+
+    def test_num_returns_streaming_with_non_generator_raises_error(self):
+        """Test that num_returns='streaming' with non-generator raises ValueError."""
+        with pytest.raises(
+            ValueError, match="num_returns='streaming' can only be used with generator"
+        ):
+
+            @ray.remote(num_returns="streaming")
+            def f():
+                return 1
+
+    def test_num_returns_dynamic_with_non_generator_raises_error(self):
+        """Test that num_returns='dynamic' with non-generator raises ValueError."""
+        with pytest.raises(
+            ValueError, match="num_returns='dynamic' can only be used with generator"
+        ):
+
+            @ray.remote(num_returns="dynamic")
+            def f():
+                return 1
+
+    def test_num_returns_streaming_with_generator_succeeds(self):
+        """Test that num_returns='streaming' with generator function succeeds."""
+
+        @ray.remote(num_returns="streaming")
+        def generator_func():
+            for i in range(3):
+                yield i
+
+    def test_num_returns_dynamic_with_generator_succeeds(self):
+        """Test that num_returns='dynamic' with generator function succeeds."""
+
+        @ray.remote(num_returns="dynamic")
+        def generator_func():
+            for i in range(3):
+                yield i
+
+    def test_num_returns_streaming_with_async_generator_succeeds(self):
+        """Test that num_returns='streaming' with async generator function succeeds."""
+
+        @ray.remote(num_returns="streaming")
+        async def async_generator_func():
+            for i in range(3):
+                yield i
+
+    def test_num_returns_positive_integer_succeeds(self):
+        """Test that num_returns with positive integer succeeds."""
+
+        @ray.remote(num_returns=2)
+        def f():
+            return 1, 2
+
+    def test_num_returns_zero_succeeds(self):
+        """Test that num_returns=0 succeeds."""
+
+        @ray.remote(num_returns=0)
+        def f():
+            return
+
+    def test_num_returns_none_succeeds(self):
+        """Test that num_returns=None succeeds."""
+
+        @ray.remote(num_returns=None)
+        def f():
+            return 1
+
+    def test_num_returns_default_succeeds(self):
+        """Test that default num_returns (not specified) succeeds."""
+
+        @ray.remote
+        def f():
+            return 1
 
 
 class TestMethodNumReturns:
@@ -10,7 +96,6 @@ class TestMethodNumReturns:
 
     def test_num_returns_negative_raises_error(self, shutdown_only):
         """Test that num_returns < 0 raises ValueError at decoration time."""
-        # validate_num_returns checks for negative values and raises an error
         with pytest.raises(ValueError, match="num_returns must be >= 0"):
 
             @ray.remote
@@ -45,7 +130,7 @@ class TestMethodNumReturns:
 
     def test_num_returns_streaming_with_generator_succeeds(self, shutdown_only):
         """Test that num_returns='streaming' with generator method succeeds."""
-        # This should not raise an error at decoration time
+
         @ray.remote
         class TestActor:
             @ray.method(num_returns="streaming")
@@ -53,12 +138,9 @@ class TestMethodNumReturns:
                 for i in range(3):
                     yield i
 
-        # Verify the class was created (no error was raised)
-        assert TestActor is not None
-
     def test_num_returns_dynamic_with_generator_succeeds(self, shutdown_only):
         """Test that num_returns='dynamic' with generator method succeeds."""
-        # This should not raise an error at decoration time
+
         @ray.remote
         class TestActor:
             @ray.method(num_returns="dynamic")
@@ -66,21 +148,15 @@ class TestMethodNumReturns:
                 for i in range(3):
                     yield i
 
-        # Verify the class was created (no error was raised)
-        assert TestActor is not None
-
     def test_num_returns_streaming_with_async_generator_succeeds(self, shutdown_only):
         """Test that num_returns='streaming' with async generator method succeeds."""
-        # This should not raise an error at decoration time
+
         @ray.remote
         class TestActor:
             @ray.method(num_returns="streaming")
             async def async_generator_method(self):
                 for i in range(3):
                     yield i
-
-        # Verify the class was created (no error was raised)
-        assert TestActor is not None
 
     def test_num_returns_positive_integer_succeeds(self, shutdown_only):
         """Test that num_returns with positive integer succeeds."""
@@ -91,9 +167,6 @@ class TestMethodNumReturns:
             def method(self):
                 return 1, 2
 
-        # Verify the class was created (no error was raised)
-        assert TestActor is not None
-
     def test_num_returns_zero_succeeds(self, shutdown_only):
         """Test that num_returns=0 succeeds."""
 
@@ -102,9 +175,6 @@ class TestMethodNumReturns:
             @ray.method(num_returns=0)
             def method(self):
                 return
-
-        # Verify the class was created (no error was raised)
-        assert TestActor is not None
 
     def test_num_returns_none_succeeds(self, shutdown_only):
         """Test that num_returns=None succeeds."""
@@ -115,9 +185,6 @@ class TestMethodNumReturns:
             def method(self):
                 return 1
 
-        # Verify the class was created (no error was raised)
-        assert TestActor is not None
-
     def test_num_returns_default_succeeds(self, shutdown_only):
         """Test that default num_returns (not specified) succeeds."""
 
@@ -125,9 +192,6 @@ class TestMethodNumReturns:
         class TestActor:
             def method(self):
                 return 1
-
-        # Verify the class was created (no error was raised)
-        assert TestActor is not None
 
 
 if __name__ == "__main__":
